pyproject.toml
[project]
name = "pnw-campsites"
version = "0.1.0"
description = "PNW campsite discovery and monitoring tool"
requires-python = ">=3.12"
dependencies = [
    "httpx==0.28.1",
    "python-dotenv==1.2.2",
    "pydantic==2.12.5",
    "curl_cffi==0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "respx==0.22.0",
    "ruff==0.15.7",
]
api = [
    "fastapi==0.135.1",
    "uvicorn==0.42.0",
    "PyJWT==2.12.1",
    "apscheduler==3.11.2",
    "pywebpush==2.3.0",
    "anthropic==0.86.0",
    "posthog>=3.0",
    "jinja2>=3.1",
]
notify = [
    "ntfy-wrapper>=0.2",
]
enrichment = [
    "anthropic==0.86.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/pnw_campsites/api.py" = ["B008"]  # Query() in FastAPI defaults is idiomatic
"src/pnw_campsites/routes/*.py" = ["B008"]  # Same for route modules

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/pnw_campsites"]
omit = ["*/scripts/*", "*/__main__.py"]

[tool.coverage.report]
fail_under = 70
show_missing = true
exclude_lines = ["if __name__", "pragma: no cover"]
